Update index.html
In making a calculator the first thing that has to be done is a from element. The name is given, and value is calculator. This is done by: <form name="calculator">

The next step would be to create a table <table> table rows <tr> and table data <td>

Inside the <table> tag table rows <tr> need to be defined. This is the horizonal section of the calculator and 
table data <td> which is going to contain the calculator buttons.

The first horizontal section is going to be the display screen.
The second horizontal section is going to be the first row of buttons and the + sign.
The third horizontal section is going to start with the number four and have the - sign.
The fourth horizontal section is going to start with the number seven and have the x sign.
The fifth horizontal section is going to start with the letter c and have the = and / sign.

The input element needs to be set up. 
<input type ="text" name ="display" id="display" 
disable>
The "text" is used to accommodate numbers and symbols.
The <disable> tag is used because we only want the user to be able to send text to the calculator display using the buttons provided. 

The first <td> tag is going to contain the number 1 and the second one is for number 2 and so on.
The input element type is going to be:
<td><input type= "button" name="one" value="1"
onclick="calculator .display.value += '1' "></td>
This is how the value of the number is displayed as a button.
The onclick attribute tells the web browser to display the value of the button when it is clicked.

The fifth row is different because of the = sign.
For the = sign to function, the eval (calculator.display.value) has to be added. What that does is to interpret the value of the input as javascript.